move all GameScreen files to GameScreen folder
