+++ src/GameScreen/Card.js
@@ -3,43 +3,43 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 
-import './css/Cards.css';
+import '../css/Cards.css';
 
-import apple from './images/cards/apple.jpg';
-import bajun from './images/cards/catbajun.jpg';
-import bat from './images/cards/bat.jpg';
-import bereginya from './images/cards/bitva-cardbase.jpg';
-import bogatyr from './images/cards/bitva-cardbase.jpg';
-import bowArrow from './images/cards/bitva-cardbase.jpg';
-import bulat from './images/cards/sword.jpg';
-import chemise from './images/cards/bitva-cardbase.jpg';
-import chickenLegsHut from './images/cards/izba.jpg';
-import clairvoyance from './images/cards/bitva-cardbase.jpg';
-import crown from './images/cards/bitva-cardbase.jpg';
-import dolly from './images/cards/bitva-cardbase.jpg';
-import earthquake from './images/cards/bitva-cardbase.jpg';
-import forestMushroom from './images/cards/mushrooms.jpg';
-import horsemanBlack from './images/cards/blackrider.jpg';
-import horsemanRed from './images/cards/redrider.jpg';
-import horsemanWhite from './images/cards/whiterider.jpg';
-import kikimora from './images/cards/bitva-cardbase.jpg';
-import lizard from './images/cards/lizard.jpg';
-import magicTree from './images/cards/tree.jpg';
-import magicMirror from './images/cards/bitva-cardbase.jpg';
-import malachiteBox from './images/cards/malachitebox.jpg';
-import mortar from './images/cards/bitva-cardbase.jpg';
-import plateMail from './images/cards/bitva-cardbase.jpg';
-import raven from './images/cards/raven.jpg';
-import russianOven from './images/cards/bitva-cardbase.jpg';
-import shieldLarge from './images/cards/largeshield.jpg';
-import shieldSmall from './images/cards/smallshield.jpg';
-import sivka from './images/cards/warhorse.jpg';
-import skullLantern from './images/cards/bitva-cardbase.jpg';
-import turningPotion from './images/cards/potion.jpg';
-import warhorse from './images/cards/warhorse.jpg';
-import waterDead from './images/cards/deadwater.jpg';
-import waterLiving from './images/cards/livingwater.jpg';
-import wolf from './images/cards/wolf.jpg';
+import apple from '../images/cards/apple.jpg';
+import bajun from '../images/cards/catbajun.jpg';
+import bat from '../images/cards/bat.jpg';
+import bereginya from '../images/cards/bitva-cardbase.jpg';
+import bogatyr from '../images/cards/bitva-cardbase.jpg';
+import bowArrow from '../images/cards/bitva-cardbase.jpg';
+import bulat from '../images/cards/sword.jpg';
+import chemise from '../images/cards/bitva-cardbase.jpg';
+import chickenLegsHut from '../images/cards/izba.jpg';
+import clairvoyance from '../images/cards/bitva-cardbase.jpg';
+import crown from '../images/cards/bitva-cardbase.jpg';
+import dolly from '../images/cards/bitva-cardbase.jpg';
+import earthquake from '../images/cards/bitva-cardbase.jpg';
+import forestMushroom from '../images/cards/mushrooms.jpg';
+import horsemanBlack from '../images/cards/blackrider.jpg';
+import horsemanRed from '../images/cards/redrider.jpg';
+import horsemanWhite from '../images/cards/whiterider.jpg';
+import kikimora from '../images/cards/bitva-cardbase.jpg';
+import lizard from '../images/cards/lizard.jpg';
+import magicTree from '../images/cards/tree.jpg';
+import magicMirror from '../images/cards/bitva-cardbase.jpg';
+import malachiteBox from '../images/cards/malachitebox.jpg';
+import mortar from '../images/cards/bitva-cardbase.jpg';
+import plateMail from '../images/cards/bitva-cardbase.jpg';
+import raven from '../images/cards/raven.jpg';
+import russianOven from '../images/cards/bitva-cardbase.jpg';
+import shieldLarge from '../images/cards/largeshield.jpg';
+import shieldSmall from '../images/cards/smallshield.jpg';
+import sivka from '../images/cards/warhorse.jpg';
+import skullLantern from '../images/cards/bitva-cardbase.jpg';
+import turningPotion from '../images/cards/potion.jpg';
+import warhorse from '../images/cards/warhorse.jpg';
+import waterDead from '../images/cards/deadwater.jpg';
+import waterLiving from '../images/cards/livingwater.jpg';
+import wolf from '../images/cards/wolf.jpg';
 
 const imagesCards = {
     apple,
